Resolving MSVC 2017 compiler warnings

+++ src/BabylonCpp/src/actions/directactions/play_animation_action.cpp
@@ -11,7 +11,7 @@ namespace BABYLON {
 
 PlayAnimationAction::PlayAnimationAction(unsigned int triggerOptions,
                                          const IAnimatablePtr& target,
-                                         int iFrom, int iTo, bool iLoop,
+                                         float iFrom, float iTo, bool iLoop,
                                          Condition* condition)
     : Action(triggerOptions, condition)
     , from{iFrom}
